c8y-mapper: extract operations module
