+++ crates/extensions/c8y_mapper_ext/src/operations/firmware_update.rs
