+++ crates/extensions/c8y_mapper_ext/src/tests.rs
@@ -2554,13 +2554,13 @@ pub(crate) async fn spawn_c8y_mapper_actor(
     let auth_proxy_addr = "127.0.0.1".into();
     let auth_proxy_port = 8001;
     let mut topics = C8yMapperConfig::default_internal_topic_filter(config_dir.path()).unwrap();
-    topics.add_all(crate::log_upload::log_upload_topic_filter(&mqtt_schema));
-    topics.add_all(crate::config_operations::config_snapshot_topic_filter(
+    topics.add_all(crate::operations::log_upload::log_upload_topic_filter(
         &mqtt_schema,
     ));
-    topics.add_all(crate::config_operations::config_update_topic_filter(
+    topics.add_all(crate::operations::config_snapshot::topic_filter(
         &mqtt_schema,
     ));
+    topics.add_all(crate::operations::config_update::topic_filter(&mqtt_schema));
     topics.add_all(C8yMapperConfig::default_external_topic_filter());
 
     let config = C8yMapperConfig::new(
